requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2